Revert "refactor(common): next.config.ts"
This reverts commit f92b7c85f1b0ec32fea7503560d8004e76bfa233.
